* added total runtime to CLI * added open fuel valve to CLI * fixed setTotalRunSecs which was only writing ints instead of longs

+++ asgeco3/Asgeco2API.cpp
@@ -228,7 +228,7 @@ void writeStates(char clientline[]){
             case API_SECS_TOT:
                 pch = strtok (NULL, FS);
 //                logg("web:SECS_TOT=" + String(atoi(pch)));
-                setTotalRunSecs(long(atoi(pch)));
+                setTotalRunSecs(atol(pch));
 //                logg("web:SECS_TOT=" + String(getTotalRunSecs()) + " set");
                 break;
             case API_GENON:
